chore(breaking): rewrite the exec API to embrace testing.
Before exec.Execer was a concrete type with private fields which did not allow users to mock it. After this change exec.Execer becomes an interface and we will be able to mock and test it. Also exec.Result becomes a struct which eases the creation of mocked results and overall simplifies the API.

As a side effect, immutable functions like FS or WithEnv became struct methods to let the implementor to decide how to deal with this functions.